ArchiveCommit - before reverting to prior

+++ Assets/---------------Scripts------------/---------------UI---------------/LifeBar.cs
@@ -5,17 +5,16 @@ using UnityEngine.UI;
 
 public class LifeBar : MonoBehaviour
 {
-    // the worlds most simple life bar slider script O_O - a bit buggy right now.
+    // The worlds most simple life bar slider script O_O - a bit buggy right now.
     public Slider slider;
-    public void SetMaxLife(int life)
+    public void SetMaxLife(float life)
     {
         slider.maxValue = life;
         slider.value = life;
     }
 
-    public void SetLife(int life)
+    public void SetLife(float life)
     {
         slider.value = life;
     }
-
 }
